+++ CHANGELOG.md
@@ -1,11 +1,32 @@
 ## Change Log
+
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## v1.1.0 - 6th June 2018 (unreleased)
+## v1.3.0 - 14th June 2018 (unreleased)
+
 ### Added
-- SDB-42 - Route provider
-- SDB-44 - Back end connection
-- SDB-34 - Submit a registration
-- SDB-49 - Establishment address (without info drop down)
-- SDB-58 - Landing page
+
+* SDB-35 - Operator Name
+
+## v1.2.1 - 13th June 2018 (unreleased)
+
+### Modified
+
+* Fixed establishment trading name info typo
+
+## v1.2.0 - 13th June 2018 (unreleased)
+
+### Added
+
+* SDB-4 - Establishment trading name
+
+## v1.1.0 - 6th June 2018 (unreleased)
+
+### Added
+
+* SDB-42 - Route provider
+* SDB-44 - Back end connection
+* SDB-34 - Submit a registration
+* SDB-49 - Establishment address (without info drop down)
+* SDB-58 - Landing page